Fix TypeError in assert.isNodeFilter when ran in Chrome
Chrome fails the test in that it returns null when accessing
the 'filter' property of a TreeWalker object (instance of
an instance of NodeFilter). While the test failing is one thing,
the test was assuming there at least be some kind of value that
can be cast to an object. Extending logic in the assertion method
to check the value before trying to assert the constructor property.