TreeWalker#null should be null if no filter is provided
Per WHATWG DOM standard IDL, the property is optional.

> interface TreeWalker {
>   attribute NodeFilter? filter;
}

And in the createTreeWalker specification, there is no statement
about it being casted into an all-accepting NodeFilter.

Rather, the filter node logic should check for the presence of a filter.

This makes it match Chrome native behaviour, allowing the unit
tests to pass both on Chrome without the polyfill.